Clarify what 'created at' time represents
During an incident, I assumed this was 'deployed at', so I didn't assume a certain commit could be to blame, but I should have.

It makes sense that it is 'created at' since it's shown before the process is complete, but we don't always think that thoroughly during an incident :)

This clarification might help.